# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SEO Automation App API
  description: |
    Comprehensive API documentation for the SEO Automation App.
    
    ## Overview
    This API provides endpoints for SEO content generation, analysis, and optimization.
    All endpoints require authentication unless otherwise specified.
    
    ## Authentication
    The API uses JWT-based authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - General endpoints: 100 requests per minute
    - AI generation endpoints: 10 requests per minute
    - Analysis endpoints: 50 requests per minute
    
  version: 1.0.0
  contact:
    name: SEO Automation Support
    email: support@seo-automation.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://seo-automation-app.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /content/generate:
    post:
      tags:
        - Content Generation
      summary: Generate SEO-optimized content
      description: |
        Generates AI-powered content with advanced SEO optimization.
        Includes competitor analysis, keyword integration, and E-E-A-T principles.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentGenerationRequest'
      responses:
        '200':
          description: Content generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentGenerationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
  /serp/analyze:
    post:
      tags:
        - SEO Analysis
      summary: Analyze SERP for keywords
      description: |
        Performs comprehensive SERP analysis including:
        - Top 10 organic results
        - Featured snippets
        - People Also Ask
        - Related searches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - keyword
                - location
              properties:
                keyword:
                  type: string
                  example: "SEO best practices"
                location:
                  type: string
                  example: "United States"
                numResults:
                  type: integer
                  default: 10
                  minimum: 1
                  maximum: 100
      responses:
        '200':
          description: SERP analysis complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SERPAnalysisResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ContentGenerationRequest:
      type: object
      required:
        - keyword
        - industry
        - targetAudience
        - wordCount
      properties:
        keyword:
          type: string
          maxLength: 100
          example: "SEO content optimization"
        industry:
          type: string
          maxLength: 50
          example: "Digital Marketing"
        targetAudience:
          type: string
          maxLength: 50
          example: "Marketing professionals"
        tone:
          type: string
          enum: [authoritative, conversational, academic, practical, thought-provoking]
          default: authoritative
        wordCount:
          type: integer
          minimum: 300
          maximum: 10000
          default: 1500
        competitorInsights:
          type: string
          description: Competitor analysis data
        targetKeywordDensity:
          type: number
          minimum: 0
          maximum: 5
          example: 2.5
        lsiKeywords:
          type: array
          items:
            type: string
          example: ["search engine", "content marketing", "keyword research"]
        entities:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum: [PERSON, ORGANIZATION, LOCATION, PRODUCT, EVENT]
    ContentGenerationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            content:
              type: string
              description: Generated content in markdown format
            wordCount:
              type: integer
            qualityAnalysis:
              type: object
              properties:
                overallScore:
                  type: number
                  minimum: 0
                  maximum: 100
                readabilityScore:
                  type: number
                grammarScore:
                  type: number
                seoScore:
                  type: number
            humanWritingAnalysis:
              type: object
              properties:
                humanScore:
                  type: number
                  minimum: 0
                  maximum: 100
                patterns:
                  type: array
                  items:
                    type: string
            eeatOptimization:
              type: object
              properties:
                experience:
                  type: number
                expertise:
                  type: number
                authoritativeness:
                  type: number
                trustworthiness:
                  type: number
        metadata:
          type: object
          properties:
            processingTime:
              type: integer
              description: Processing time in milliseconds
            model:
              type: string
              example: "gpt-4o"
            version:
              type: string
...
    SERPAnalysisResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            keyword:
              type: string
            location:
              type: string
            searchVolume:
              type: integer
            difficulty:
              type: number
            organicResults:
              type: array
              items:
                type: object
                properties:
                  position:
                    type: integer
                  title:
                    type: string
                  url:
                    type: string
                  snippet:
                    type: string
                  domain:
                    type: string
            featuredSnippet:
              type: object
              nullable: true
            peopleAlsoAsk:
              type: array
              items:
                type: string
            relatedSearches:
              type: array
              items:
                type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Validation Error"
            message: "Invalid request parameters"
            details:
              keyword: "Keyword is required"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
            message: "Authentication required"
...
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate Limit Exceeded"
            message: "Too many requests. Please try again later."
            details:
              retryAfter: 60
...